Table layout for unused entities
