+++ src/panels/lovelace/editor/select-view/show-select-view-dialog.ts
@@ -1,0 +1,26 @@
+import { fireEvent } from "../../../../common/dom/fire_event";
+import { Lovelace } from "../../types";
+
+declare global {
+  // for fire event
+  interface HASSDomEvents {
+    "show-select-view": SelectViewDialogParams;
+  }
+}
+
+export interface SelectViewDialogParams {
+  lovelace: Lovelace;
+  viewSelectedCallback: (view: number) => void;
+}
+
+export const showSelectViewDialog = (
+  element: HTMLElement,
+  selectViewDialogParams: SelectViewDialogParams
+): void => {
+  fireEvent(element, "show-dialog", {
+    dialogTag: "hui-dialog-select-view",
+    dialogImport: () =>
+      import(/* webpackChunkName: "hui-dialog-select-view" */ "./hui-dialog-select-view"),
+    dialogParams: selectViewDialogParams,
+  });
+};
